// fe/src/lib/api/generated/model/mediaWorkspaceState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { RecordingWorkspaceState } from './recordingWorkspaceState';
import type { TranscriptionWorkspaceState } from './transcriptionWorkspaceState';

export interface MediaWorkspaceState {
  recording: RecordingWorkspaceState;
  transcription: TranscriptionWorkspaceState;
}